Feature/zip to dll migration (#494)
* Additional directory check

* Fixed comments, also made sure DLL is downloaded before attempting to export it

* When deleting new DLL based record, delete its parent directory if it is empty

* When downloading zip, extract dll and then delete zip

* Migrate DLLs from zip to dlls on disk

* Remove background of info bar when downloading DLL

* Change from extracting zip to swapping DLLs

This needs testing before merging to main

* Additional note about use of firstordefault

* test comments

* test commit

* revert test commits

* Fix issue where updating manifest does not actually update manifest

* Fix issue where XeSS records are tagged DLSS_D

* Don't save manifest if it matches the previously downloaded data

* Fixed export for new file structure

* Update migrating system to display that it is migrating

Also changes migration code to not run if there is no zip folders.

* Actually copy the DLLs, this is the entire point of DLSS Swapper

+++ src/Data/LocalRecord.cs
@@ -59,7 +59,21 @@ public partial class LocalRecord : ObservableObject, IEquatable<LocalRecord>
     {
         try
         {
-            File.Delete(ExpectedPath);
+            if (File.Exists(ExpectedPath))
+            {
+                File.Delete(ExpectedPath);
+            }
+
+            var path = Path.GetDirectoryName(ExpectedPath) ?? string.Empty;
+            if (string.IsNullOrWhiteSpace(path) == false)
+            {
+                if (Directory.GetFiles(path).Length == 0 && Directory.GetDirectories(path).Length == 0)
+                {
+                    Directory.Delete(path);
+                }
+            }
+
+
 
             IsDownloaded = false;
             HasDownloadError = false;
